+++ src/index.js
@@ -2,31 +2,12 @@
 
 import Store from './modules/storage.js';
 import User from './modules/Users.js';
+import UI from './modules/UIclass.js';
 
 // const name = document.querySelector('#username');
 // const score = document.querySelector('#score').value;
 const form = document.querySelector('#form');
 // const userInput = [];
-
-class UI {
-  static displayUserScore() {
-    const userInput = Store.getUser();
-    userInput.forEach((user) => UI.addUserToList(user));
-  }
-
-  static addUserToList(newUser) {
-    // console.log(newUser)
-    const list = document.querySelector('.List');
-    const row = document.createElement('tr');
-    row.innerHTML = `<td>${newUser.score}: ${newUser.name}</td>`;
-    list.appendChild(row);
-  }
-
-  static clearFields() {
-    document.querySelector('#username').value = '';
-    document.querySelector('#score').value = '';
-  }
-}
 
 // Event : Display UsersScore
 document.addEventListener('DOMcontentLoaded', UI.displayUserScore());
